requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.5.0
dnspython==2.4.2
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
gunicorn==21.2.0
pandas==2.0.3
